requirements.txt
beautifulsoup4==4.12.3
Django==5.0.6
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-utils2==3.0.2
gunicorn==22.0.0
json-five==1.1.2
python-dateutil==2.9.0.post0
urllib3==2.2.2
whitenoise==6.7.0
django-user-agents==0.4.0
requests==2.32.3
django-cleanup==8.1.0
Pillow==10.3.0
djangorestframework==3.15.2
razorpay==1.4.2
django-storages==1.14.3
django-meta==2.4.2
django-environ==0.11.2
razorpay==1.4.2